# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: HAS SmartLock Service API
  version: 0.1.0
  description: |
    当前仓库已实现接口的 OpenAPI 定义。
    用于导入 Apifox、Postman 等 API 测试平台。
servers:
  - url: http://localhost:8080
    description: Local development
...
paths:
...
  /v1/user/registerSend:
    post:
      tags:
        - User
      summary: 发送注册验证码
      operationId: userRegisterSend
      description: |
        用户端常规签名接口。需携带 appid、app_version、phone_code、timestamp、request_id、sign 等 Header。
      parameters:
        - $ref: '#/components/parameters/AppIDHeader'
        - $ref: '#/components/parameters/AppVersionHeader'
        - $ref: '#/components/parameters/PhoneCodeHeader'
        - $ref: '#/components/parameters/TimestampHeader'
        - $ref: '#/components/parameters/RequestIDHeader'
        - $ref: '#/components/parameters/SignHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SendCodeRequest'
            example:
              username: '{{username}}'
              country: '{{country}}'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NullEnvelope'
              example:
                code: 1000
                msg: ok
                data: null
  /v1/user/loginSend:
    post:
      tags:
        - User
      summary: 发送登录验证码
      operationId: userLoginSend
      description: |
        用户端常规签名接口。需携带 appid、app_version、phone_code、timestamp、request_id、sign 等 Header。
      parameters:
        - $ref: '#/components/parameters/AppIDHeader'
        - $ref: '#/components/parameters/AppVersionHeader'
        - $ref: '#/components/parameters/PhoneCodeHeader'
        - $ref: '#/components/parameters/TimestampHeader'
        - $ref: '#/components/parameters/RequestIDHeader'
        - $ref: '#/components/parameters/SignHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SendCodeRequest'
            example:
              username: '{{username}}'
              country: '{{country}}'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NullEnvelope'
              example:
                code: 1000
                msg: ok
                data: null
  /v1/user/register:
    post:
      tags:
        - User
      summary: 用户注册
      operationId: userRegister
      description: |
        用户端常规签名接口。需携带 appid、app_version、phone_code、timestamp、request_id、sign 等 Header。
      parameters:
        - $ref: '#/components/parameters/AppIDHeader'
        - $ref: '#/components/parameters/AppVersionHeader'
        - $ref: '#/components/parameters/PhoneCodeHeader'
        - $ref: '#/components/parameters/TimestampHeader'
        - $ref: '#/components/parameters/RequestIDHeader'
        - $ref: '#/components/parameters/SignHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
            example:
              username: '{{username}}'
              country: '{{country}}'
              code: '{{code}}'
              password: '{{password}}'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenEnvelope'
              example:
                code: 1000
                msg: ok
                data:
                  uid: u_123456789
                  access_token: access-token
                  refresh_token: refresh-token
                  is_debug: 0
                  expiration: 1711864800
  /v1/user/login:
    post:
      tags:
        - User
      summary: 用户登录
      operationId: userLogin
      description: |
        用户端常规签名接口。需携带 appid、app_version、phone_code、timestamp、request_id、sign 等 Header。
      parameters:
        - $ref: '#/components/parameters/AppIDHeader'
        - $ref: '#/components/parameters/AppVersionHeader'
        - $ref: '#/components/parameters/PhoneCodeHeader'
        - $ref: '#/components/parameters/TimestampHeader'
        - $ref: '#/components/parameters/RequestIDHeader'
        - $ref: '#/components/parameters/SignHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            examples:
              passwordLogin:
                summary: 密码登录
                value:
                  username: '{{username}}'
                  type: password
                  password: '{{password}}'
                  phone_brand: '{{phone_brand}}'
              codeLogin:
                summary: 验证码登录
                value:
                  username: '{{username}}'
                  type: code
                  code: '{{code}}'
                  phone_brand: '{{phone_brand}}'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenEnvelope'
  /v1/user/validateCode:
    post:
      tags:
        - User
      summary: 检查验证码
      operationId: userValidateCode
      description: |
        用户端常规签名接口。需携带 appid、app_version、phone_code、timestamp、request_id、sign 等 Header。
      parameters:
        - $ref: '#/components/parameters/AppIDHeader'
        - $ref: '#/components/parameters/AppVersionHeader'
        - $ref: '#/components/parameters/PhoneCodeHeader'
        - $ref: '#/components/parameters/TimestampHeader'
        - $ref: '#/components/parameters/RequestIDHeader'
        - $ref: '#/components/parameters/SignHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ValidateCodeRequest'
            example:
              username: '{{username}}'
              code: '{{code}}'
              type: register
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NullEnvelope'
  /v1/user/resetSend:
    post:
      tags:
        - User
      summary: 发送忘记密码验证码
      operationId: userResetSend
      description: |
        用户端常规签名接口。需携带 appid、app_version、phone_code、timestamp、request_id、sign 等 Header。
      parameters:
        - $ref: '#/components/parameters/AppIDHeader'
        - $ref: '#/components/parameters/AppVersionHeader'
        - $ref: '#/components/parameters/PhoneCodeHeader'
        - $ref: '#/components/parameters/TimestampHeader'
        - $ref: '#/components/parameters/RequestIDHeader'
        - $ref: '#/components/parameters/SignHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SendCodeRequest'
            example:
              username: '{{username}}'
              country: '{{country}}'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NullEnvelope'
  /v1/user/reset:
    post:
      tags:
        - User
      summary: 忘记密码（重置密码）
      operationId: userReset
      description: |
        用户端常规签名接口。需携带 appid、app_version、phone_code、timestamp、request_id、sign 等 Header。
      parameters:
        - $ref: '#/components/parameters/AppIDHeader'
        - $ref: '#/components/parameters/AppVersionHeader'
        - $ref: '#/components/parameters/PhoneCodeHeader'
        - $ref: '#/components/parameters/TimestampHeader'
        - $ref: '#/components/parameters/RequestIDHeader'
        - $ref: '#/components/parameters/SignHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResetRequest'
            example:
              username: '{{username}}'
              code: '{{code}}'
              password: '{{new_password}}'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NullEnvelope'
  /v1/user/refresh:
    post:
      tags:
        - User
      summary: 刷新 Token
      operationId: userRefresh
      description: |
        用户端常规签名接口。需携带 appid、app_version、phone_code、timestamp、request_id、sign 等 Header。
      parameters:
        - $ref: '#/components/parameters/AppIDHeader'
        - $ref: '#/components/parameters/AppVersionHeader'
        - $ref: '#/components/parameters/PhoneCodeHeader'
        - $ref: '#/components/parameters/TimestampHeader'
        - $ref: '#/components/parameters/RequestIDHeader'
        - $ref: '#/components/parameters/SignHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
            example:
              refresh_token: '{{refresh_token}}'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenEnvelope'
  /v1/user/info:
    get:
      tags:
        - User
      summary: 获取用户信息
      operationId: userInfo
      description: |
        用户端常规签名接口。除签名头外，Auth 为必需，因此还需要 access_token Header。
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/AppIDHeader'
        - $ref: '#/components/parameters/AppVersionHeader'
        - $ref: '#/components/parameters/PhoneCodeHeader'
        - $ref: '#/components/parameters/TimestampHeader'
        - $ref: '#/components/parameters/RequestIDHeader'
        - $ref: '#/components/parameters/SignHeader'
        - $ref: '#/components/parameters/AccessTokenHeader'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserInfoEnvelope'
              example:
                code: 1000
                msg: ok
                data:
                  username: demo@example.com
                  nickname: demo@example.com
                  avatar: avatar/u_123456789
                  is_debug: 0
                  register_time: 1711857600
  /v1/user/logout:
    post:
      tags:
        - User
      summary: 退出登录
      operationId: userLogout
      description: |
        用户端常规签名接口。除签名头外，Auth 为必需，因此还需要 access_token Header。
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/AppIDHeader'
        - $ref: '#/components/parameters/AppVersionHeader'
        - $ref: '#/components/parameters/PhoneCodeHeader'
        - $ref: '#/components/parameters/TimestampHeader'
        - $ref: '#/components/parameters/RequestIDHeader'
        - $ref: '#/components/parameters/SignHeader'
        - $ref: '#/components/parameters/AccessTokenHeader'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NullEnvelope'
  /v1/user/updatePwd:
    post:
      tags:
        - User
      summary: 修改密码
      operationId: userUpdatePwd
      description: |
        用户端常规签名接口。除签名头外，Auth 为必需，因此还需要 access_token Header。
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/AppIDHeader'
        - $ref: '#/components/parameters/AppVersionHeader'
        - $ref: '#/components/parameters/PhoneCodeHeader'
        - $ref: '#/components/parameters/TimestampHeader'
        - $ref: '#/components/parameters/RequestIDHeader'
        - $ref: '#/components/parameters/SignHeader'
        - $ref: '#/components/parameters/AccessTokenHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePasswordRequest'
            example:
              new_password: '{{new_password}}'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NullEnvelope'
  /v1/user/updateInfo:
    post:
      tags:
        - User
      summary: 修改昵称
      operationId: userUpdateInfo
      description: |
        用户端常规签名接口。除签名头外，Auth 为必需，因此还需要 access_token Header。
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/AppIDHeader'
        - $ref: '#/components/parameters/AppVersionHeader'
        - $ref: '#/components/parameters/PhoneCodeHeader'
        - $ref: '#/components/parameters/TimestampHeader'
        - $ref: '#/components/parameters/RequestIDHeader'
        - $ref: '#/components/parameters/SignHeader'
        - $ref: '#/components/parameters/AccessTokenHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateInfoRequest'
            example:
              nickname: '{{nickname}}'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NullEnvelope'
  /v1/user/putClient:
    post:
      tags:
        - User
      summary: 上报登录设备信息
      operationId: userPutClient
      description: |
        用户端常规签名接口。除签名头外，Auth 为必需，因此还需要 access_token Header。
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/AppIDHeader'
        - $ref: '#/components/parameters/AppVersionHeader'
        - $ref: '#/components/parameters/PhoneCodeHeader'
        - $ref: '#/components/parameters/TimestampHeader'
        - $ref: '#/components/parameters/RequestIDHeader'
        - $ref: '#/components/parameters/SignHeader'
        - $ref: '#/components/parameters/AccessTokenHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PutClientRequest'
            example:
              push_type: 1
              push_token: '{{push_token}}'
              brand: '{{brand}}'
              version: '{{version}}'
              language: '{{language}}'
              zone: '{{zone}}'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NullEnvelope'
...
  /v1/user/deleteSend:
    post:
      tags:
        - User
      summary: 发送删除帐号验证码
      operationId: userDeleteSend
      description: |
        用户端常规签名接口。除签名头外，Auth 为必需，因此还需要 access_token Header。
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/AppIDHeader'
        - $ref: '#/components/parameters/AppVersionHeader'
        - $ref: '#/components/parameters/PhoneCodeHeader'
        - $ref: '#/components/parameters/TimestampHeader'
        - $ref: '#/components/parameters/RequestIDHeader'
        - $ref: '#/components/parameters/SignHeader'
        - $ref: '#/components/parameters/AccessTokenHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeleteSendRequest'
            example:
              username: '{{username}}'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NullEnvelope'
  /v1/user/delete:
    post:
      tags:
        - User
      summary: 删除帐号
      operationId: userDelete
      description: |
        用户端常规签名接口。除签名头外，Auth 为必需，因此还需要 access_token Header。
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/AppIDHeader'
        - $ref: '#/components/parameters/AppVersionHeader'
        - $ref: '#/components/parameters/PhoneCodeHeader'
        - $ref: '#/components/parameters/TimestampHeader'
        - $ref: '#/components/parameters/RequestIDHeader'
        - $ref: '#/components/parameters/SignHeader'
        - $ref: '#/components/parameters/AccessTokenHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeleteRequest'
            example:
              username: '{{username}}'
              code: '{{code}}'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NullEnvelope'
components:
  parameters:
    AppIDHeader:
      in: header
      name: appid
      required: true
      example: '{{appid}}'
      schema:
        type: string
        default: '{{appid}}'
      description: 应用 ID
    AppVersionHeader:
      in: header
      name: app_version
      required: true
      example: '{{app_version}}'
      schema:
        type: string
        default: '{{app_version}}'
      description: 应用版本号
    PhoneCodeHeader:
      in: header
      name: phone_code
      required: true
      example: '{{phone_code}}'
      schema:
        type: string
        default: '{{phone_code}}'
      description: 手机国家区号
    TimestampHeader:
      in: header
      name: timestamp
      required: true
      example: '{{timestamp}}'
      schema:
        type: string
        default: '{{timestamp}}'
      description: Unix 秒级时间戳
    RequestIDHeader:
      in: header
      name: request_id
      required: true
      example: '{{request_id}}'
      schema:
        type: string
        default: '{{request_id}}'
      description: 请求唯一 ID
    SignHeader:
      in: header
      name: sign
      required: true
      example: '{{sign}}'
      schema:
        type: string
        default: '{{sign}}'
      description: 请求签名
    AccessTokenHeader:
      in: header
      name: access_token
      required: true
      example: '{{access_token}}'
      schema:
        type: string
        default: '{{access_token}}'
      description: 访问令牌
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    EnvelopeBase:
      type: object
      required:
        - code
        - msg
      properties:
        code:
          type: integer
          example: 1000
        msg:
          type: string
          example: ok
...
    NullEnvelope:
      allOf:
        - $ref: '#/components/schemas/EnvelopeBase'
        - type: object
          properties:
            data:
              nullable: true
    SendCodeRequest:
      type: object
      required:
        - username
        - country
      properties:
        username:
          type: string
          example: '{{username}}'
        country:
          type: string
          example: '{{country}}'
    RegisterRequest:
      type: object
      required:
        - username
        - country
        - code
      properties:
        username:
          type: string
        country:
          type: string
        code:
          type: string
        password:
          type: string
          nullable: true
    LoginRequest:
      type: object
      required:
        - username
        - type
        - phone_brand
      properties:
        username:
          type: string
        type:
          type: string
          enum:
            - password
            - code
        password:
          type: string
          nullable: true
        code:
          type: string
          nullable: true
        phone_brand:
          type: string
      description: 当前实现支持密码登录和验证码登录，字段约束以后续主合同文档为准。
    ValidateCodeRequest:
      type: object
      required:
        - username
        - code
        - type
      properties:
        username:
          type: string
        code:
          type: string
        type:
          type: string
          enum:
            - login
            - register
            - reset
    RefreshRequest:
      type: object
      required:
        - refresh_token
      properties:
        refresh_token:
          type: string
    ResetRequest:
      type: object
      required:
        - username
        - code
        - password
      properties:
        username:
          type: string
        code:
          type: string
        password:
          type: string
    UpdatePasswordRequest:
      type: object
      required:
        - new_password
      properties:
        new_password:
          type: string
    UpdateInfoRequest:
      type: object
      required:
        - nickname
      properties:
        nickname:
          type: string
    PutClientRequest:
      type: object
      required:
        - push_type
        - push_token
      properties:
        push_type:
          type: integer
        push_token:
          type: string
        brand:
          type: string
        version:
          type: string
        language:
          type: string
        zone:
          type: string
    DeleteSendRequest:
      type: object
      required:
        - username
      properties:
        username:
          type: string
    DeleteRequest:
      type: object
      required:
        - username
        - code
      properties:
        username:
          type: string
        code:
          type: string
...
    TokenData:
      type: object
      required:
        - uid
        - access_token
        - refresh_token
        - is_debug
        - expiration
      properties:
        uid:
          type: string
        access_token:
          type: string
        refresh_token:
          type: string
        is_debug:
          type: integer
        expiration:
          type: integer
          format: int64
    TokenEnvelope:
      allOf:
        - $ref: '#/components/schemas/EnvelopeBase'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/TokenData'
    UserInfoData:
      type: object
      required:
        - username
        - nickname
        - avatar
        - is_debug
        - register_time
      properties:
        username:
          type: string
        nickname:
          type: string
        avatar:
          type: string
        is_debug:
          type: integer
        register_time:
          type: integer
          format: int64
    UserInfoEnvelope:
      allOf:
        - $ref: '#/components/schemas/EnvelopeBase'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/UserInfoData'
...